refactor: split entities.ts and list.ts god files into focused modules (#997)
Split two large "god files" into focused modules to improve maintainability:

## entities.ts (2676 lines → 4 modules)

Created src/core/entities/ subdirectory:
- schemas.ts (839 lines) - All Zod config schemas + types
- helpers.ts (413 lines) - All helper functions
- factories.ts (769 lines) - All 15 entity factory functions
- index.ts (59 lines) - Re-export hub

Original entities.ts reduced to 57 lines (backward compat re-export).

## list.ts (2180 lines → 18 modules)

Created src/components/list/ subdirectory with 17 focused modules:
- types.ts - Type definitions and interfaces
- constants.ts - Constants and defaults
- stores.ts - Data stores and maps
- stateMachine.ts - State machine configuration
- core.ts - Core list behavior and lifecycle
- items.ts - Item CRUD operations
- selection.ts - Selection management
- callbacks.ts - Event callbacks
- multiSelect.ts - Multi-select mode
- keyHandling.ts - Key input handling
- virtualization.ts - Viewport and lazy loading
- display.ts - Display configuration
- options.ts - Options management
- rendering.ts - Rendering helpers
- search.ts - Search mode functionality
- filter.ts - Filtering functionality
- reset.ts - Testing utilities
- index.ts - Re-export hub

Original list.ts reduced to 34 lines (backward compat re-export).

## Verification

- ✅ 100% backward compatibility maintained
- ✅ All 10,986 tests passing
- ✅ TypeScript compilation successful
- ✅ Lint passes (13 warnings, 0 errors)
- ✅ Build successful
