+++ src/components/list/constants.ts
@@ -1,0 +1,32 @@
+/**
+ * List Component Constants
+ *
+ * @module components/list/constants
+ */
+
+/** Default selected prefix character */
+export const DEFAULT_SELECTED_PREFIX = '> ';
+
+/** Default unselected prefix character */
+export const DEFAULT_UNSELECTED_PREFIX = '  ';
+
+/** Default selected foreground color */
+export const DEFAULT_SELECTED_FG = 0xffffffff;
+
+/** Default selected background color */
+export const DEFAULT_SELECTED_BG = 0x0066ffff;
+
+/** Default item foreground color */
+export const DEFAULT_ITEM_FG = 0xccccccff;
+
+/** Default item background color */
+export const DEFAULT_ITEM_BG = 0x000000ff;
+
+/** Default disabled foreground color */
+export const DEFAULT_DISABLED_FG = 0x666666ff;
+
+/** Maximum entities supported */
+export const MAX_ENTITIES = 10000;
+
+/** Default loading placeholder */
+export const DEFAULT_LOADING_PLACEHOLDER = 'Loading...';
