+++ src/core/entities/index.ts
@@ -1,0 +1,59 @@
+/**
+ * Entity factory functions and configuration schemas.
+ * This module provides functions for creating common UI entities with pre-configured components.
+ * @module core/entities
+ */
+
+// Re-export all factory functions
+export {
+	createBoxEntity,
+	createButtonEntity,
+	createCheckboxEntity,
+	createFormEntity,
+	createInputEntity,
+	createListEntity,
+	createProgressBarEntity,
+	createRadioButtonEntity,
+	createRadioSetEntity,
+	createScreenEntity,
+	createSelectEntity,
+	createSliderEntity,
+	createTextareaEntity,
+	createTextboxEntity,
+	createTextEntity,
+} from './factories';
+// Re-export all schemas and types
+export type {
+	BoxConfig,
+	ButtonConfig,
+	CheckboxConfig,
+	FormConfig,
+	InputConfig,
+	ListConfig,
+	ProgressBarConfig,
+	RadioButtonConfig,
+	RadioSetConfig,
+	ScreenConfig,
+	SelectConfig,
+	SliderConfig,
+	TextareaConfig,
+	TextboxConfig,
+	TextConfig,
+} from './schemas';
+export {
+	BoxConfigSchema,
+	ButtonConfigSchema,
+	CheckboxConfigSchema,
+	FormConfigSchema,
+	InputConfigSchema,
+	ListConfigSchema,
+	ProgressBarConfigSchema,
+	RadioButtonConfigSchema,
+	RadioSetConfigSchema,
+	ScreenConfigSchema,
+	SelectConfigSchema,
+	SliderConfigSchema,
+	TextareaConfigSchema,
+	TextboxConfigSchema,
+	TextConfigSchema,
+} from './schemas';
